dom/tests/mochitest/bugs/test_bug1551425_reference_to_childnodes_nodelist_keeps_childnodes_alive.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1551425
-->
<head>
  <meta charset="utf-8">
  <title>
      Test for bug 1551425: check reference to childNodes' NodeList keeps the
      child nodes alive.
  </title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1551425">Mozilla bug 1551425</a>
<p id="display"></p>
<div id="content" style="display: none">
  <div id="d"><div>d2</div></div>
</div>

<pre id="test">
  <script>
    var d = document.getElementById("d");

    // Get childNodes, don't keep reference to parent.
    var childNodes = d.cloneNode(true).childNodes;

    // Trigger GC to remove the cloned node.
    SpecialPowers.gc();

    is(childNodes.length, 1);
    is(childNodes[0].textContent, "d2");
  </script>
</pre>
</body>
</html>
